fix pulse counter multiplier to support floats

+++ insighioNode/apps/demo_console/scenario_pcnt_ulp.py
@@ -234,7 +234,7 @@ def read_ulp_values(measurements, multiplier, label=None):
 
     pcnt_multiplier = 1
     try:
-        pcnt_multiplier = int(multiplier)
+        pcnt_multiplier = float(multiplier)
     except:
         pass
